requirements.txt
DOCKER
fastapi
uvicorn
presidio-analyzer
pathwaycom
python-dotenv
sentence-transformers